refactor(TextArea): Use `partials` pattern
