+++ packages/react-component-library/src/components/TextArea/partials/StyledTextArea.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const StyledTextArea = styled.div`
+  display: inline-flex;
+  flex-direction: column;
+  position: relative;
+  padding: 0;
+  border: 0;
+  vertical-align: top;
+  width: 100%;
+`
