+++ packages/react-component-library/src/components/TextArea/partials/StyledTextAreaWrapperProps.tsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+import { selectors } from '@royalnavy/design-tokens'
+
+import {
+  getOuterWrapperBorder,
+  StyledOuterWrapperProps,
+} from '../../../styled-components'
+
+const { color } = selectors
+
+export const StyledTextAreaWrapper = styled.div<StyledOuterWrapperProps>`
+  position: relative;
+  background: ${color('neutral', 'white')};
+
+  ${(props) => getOuterWrapperBorder(props)}
+`
